+++ TireFire/Public/Set-TireFireConfig.ps1
@@ -1,11 +1,15 @@
 function Set-TireFireConfig {
     <#
     .SYNOPSIS
-       Set TireFire configuration values
+        Set TireFire configuration values
     .DESCRIPTION
-       Set TireFire configuration values
+        Set TireFire configuration values
     .EXAMPLE
-        Set-TireFireConfig -Token $Token
+        Set-TireFireConfig -Backend File -BackendConfig @{RootPath = 'C:\notes'}
+        Use a specific backend and configure it
+    .EXAMPLE
+        Set-TireFireConfig -BackendConfig @{RootPath = '\\path\to\notes'}
+        Updated the path to notes
     .PARAMETER Backend
         Backend to use for all commands
     .PARAMETER BackendConfig