+++ CHANGELOG.md
@@ -5,4 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1] Unreleased
+## [0.0.3]
+
+* Add -ComparisonOperator to Get-Note - Or returns faster, And is more selective
+* Add -MergeData to Get-Note.  Shifts all toplevel properties from Data into the note itself
+* Bug fixes and things
+* Apparently tests are a thing?
+
+## [0.0.2]
+
+* Bug fixes and things
+
+## [0.0.1]
+
+* Initial release
+